+++ ReleaseNotes.md
@@ -84,3 +84,4 @@
 - Add MileCreateCoreWindow function.
 - Add Mile::ComObjectChainHelper template struct.
 - Add Mile::ComObjectQueryHelper template struct.
+- Add Mile::ComObject template struct.